refactor(chain): move CHAIN INTEL into the RECON toolkit
Follows #253. The brief and wallet lookup shipped as a floating right-rail
panel; this relocates both into the RECON toolkit's CHAIN INTEL tab, which
now carries two views:

  DAILY BRIEF       exploits, CVEs, OFAC wallets, AI overview
  WALLET FORENSICS  the per-address lookup

The brief takes no target, so the toolkit's shared query bar is hidden
while it is selected and returns for the wallet view. The results
container renders without a prior lookup now, since the brief loads
itself.

The brief moves into a self-contained ChainBrief component owning its
fetch, window selector, unattended refresh and AI overview, so the host
only drops it in. ChainPanel is deleted along with its rail button and
state. Nothing is lost — the API routes, risk scoring and OFAC screening
are untouched — only relocated.
