+++ src/components/OsintPanel.tsx
@@ -12,6 +12,7 @@ import {
   Maximize2, Minimize2, Gavel, Bitcoin, Phone, Terminal, ShieldAlert
 } from 'lucide-react';
 import { ipToNumber, numberToIp, calculateSubnetStart, classifyDevice, assessRisk, batchFetch, ShodanInternetDBResponse, SweepDevice } from '@/lib/osint-utils';
+import ChainBrief from '@/components/ChainBrief';
 
 const TABS = [
   { id: 'scanner', label: 'PORT SCAN', icon: Radar, placeholder: 'IP or hostname', color: '#00E5FF' },
@@ -31,6 +32,7 @@ const TABS = [
   { id: 'phone', label: 'PHONE INTEL', icon: Phone, placeholder: 'Phone number (e.g. +1...)', color: '#FF9500' },
   { id: 'leaks', label: 'DATA LEAKS', icon: ShieldAlert, placeholder: 'Email address', color: '#E040FB' },
   { id: 'github', label: 'GITHUB RECON', icon: Terminal, placeholder: 'GitHub username', color: '#87CEEB' },
+  { id: 'crypto', label: 'CHAIN INTEL', icon: Bitcoin, placeholder: 'BTC, ETH or SOL wallet address', color: '#F7931A' },
   { id: 'sweep', label: 'IP SWEEP', icon: Crosshair, placeholder: 'Enter IP address (e.g. 8.8.8.8)', color: '#FF3D3D' },
 ];
 
@@ -50,6 +52,9 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
   const [sweepProgress, setSweepProgress] = useState<{ current: number; total: number } | null>(null);
   const [sweepCidr, setSweepCidr] = useState(24);
   const [cveCache, setCveCache] = useState<Record<string, any>>({});
+  // CHAIN INTEL carries two views: the daily brief needs no target, the wallet
+  // lookup uses the shared query bar.
+  const [chainView, setChainView] = useState<'brief' | 'wallet'>('brief');
   const [expandedDevice, setExpandedDevice] = useState<string | null>(null);
 
   // Fetch CVE details when a device is expanded in full-screen mode
@@ -329,6 +334,16 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
   );
 
   const renderStructuredResults = () => {
+    // The daily brief is self-loading and needs no query, so it renders
+    // regardless of whether a lookup has been run.
+    if (activeTab === 'crypto' && chainView === 'brief') {
+      return (
+        <div>
+          <SectionHeader title="CHAIN INTEL — DAILY BRIEF" icon={Bitcoin} color="#F7931A" />
+          <ChainBrief />
+        </div>
+      );
+    }
     if (!results) return null;
     const r = results;
 
@@ -569,6 +584,174 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
       );
     }
 
+    // ── CHAIN INTEL ──
+    // Shares /api/osint/crypto with the standalone CHAIN panel; this is the
+    // in-toolkit view of the same wallet report.
+    if (activeTab === 'crypto') {
+      const ACCENT = '#F7931A';
+      const RISK_COLOR: Record<string, string> = {
+        critical: '#FF1744', high: '#FF3D3D', medium: '#FF9500', low: '#FFD700', info: '#00E676',
+      };
+      const riskColor = RISK_COLOR[r.risk?.level] || '#00E676';
+      const fmt = (n: number, d = 4) =>
+        typeof n === 'number' && Number.isFinite(n)
+          ? n.toLocaleString(undefined, { maximumFractionDigits: d })
+          : '—';
+      const short = (a: string) => (a && a.length > 20 ? `${a.slice(0, 10)}…${a.slice(-8)}` : a);
+
+      return (
+        <div>
+          <SectionHeader title="CHAIN INTEL" icon={Bitcoin} color={ACCENT} />
+
+          {/* A sanctions hit dominates the report — surface it before anything else. */}
+          {r.sanctions?.hit && (
+            <div className="mb-2 px-2 py-2 rounded border border-red-500/40 bg-red-500/15">
+              <div className="flex items-center gap-2 mb-1.5">
+                <AlertTriangle className="w-3.5 h-3.5 text-red-400" />
+                <span className="text-[10px] font-mono font-bold text-red-400 tracking-wider">
+                  OFAC SANCTIONED WALLET
+                </span>
+              </div>
+              {(r.sanctions.entries || []).map((e: any, i: number) => (
+                <div key={i} className="text-[9px] font-mono text-red-200 break-all leading-tight">
+                  ↳ {e.name}{e.programs?.length ? ` — ${e.programs.join(', ')}` : ''}
+                </div>
+              ))}
+            </div>
+          )}
+
+          <div className="flex items-center gap-2 mb-2">
+            <span className="px-2 py-0.5 rounded text-[9px] font-mono font-bold border"
+              style={{ color: ACCENT, borderColor: `${ACCENT}55`, background: `${ACCENT}18` }}>
+              {r.chain_label?.toUpperCase()}
+            </span>
+            <span className="px-2 py-0.5 rounded text-[9px] font-mono font-bold border"
+              style={{ color: riskColor, borderColor: `${riskColor}55`, background: `${riskColor}18` }}>
+              RISK {r.risk?.score} · {String(r.risk?.level || '').toUpperCase()}
+            </span>
+          </div>
+
+          <ResultRow label="Address" value={r.address} color={ACCENT} />
+          <ResultRow label="Balance" value={`${fmt(r.balance?.native, 8)} ${r.symbol}`} color="#00E676" />
+          <ResultRow
+            label="Value (USD)"
+            value={r.balance?.usd != null ? `$${fmt(r.balance.usd, 2)}` : 'price unavailable'}
+          />
+          <ResultRow label="Spot price" value={r.balance?.price_usd != null ? `$${fmt(r.balance.price_usd, 2)}` : null} />
+          <ResultRow label="Transactions" value={r.activity?.tx_count?.toLocaleString()} />
+          <ResultRow label="Last active" value={r.activity?.last_seen ? `${String(r.activity.last_seen).slice(0, 10)} (${r.activity.dormant_days}d ago)` : null} />
+          {/* age_days is withheld by the API whenever the sample cannot prove it. */}
+          <ResultRow
+            label="Age"
+            value={r.activity?.age_days != null ? `${r.activity.age_days} days` : 'unknown — history exceeds sample'}
+            color={r.activity?.age_days != null ? undefined : 'var(--text-muted)'}
+          />
+          {r.ambiguous_chain && (
+            <ResultRow label="Note" value="Address format is valid on both Bitcoin and Solana; assumed Bitcoin." color="#FFD700" />
+          )}
+
+          {r.labels?.length > 0 && (
+            <div className="flex flex-wrap gap-1 mt-2">
+              {r.labels.map((l: string, i: number) => (
+                <span key={i} className="px-1.5 py-0.5 rounded text-[9px] font-mono border border-white/15 text-[var(--text-secondary)] bg-white/5">
+                  {l}
+                </span>
+              ))}
+            </div>
+          )}
+
+          {r.flow && (
+            <>
+              <SectionHeader title="FLOW" icon={Layers} color={ACCENT} />
+              <ResultRow label="Total in" value={`${fmt(r.flow.total_in)} ${r.symbol}`} color="#00E676" />
+              <ResultRow label="Total out" value={`${fmt(r.flow.total_out)} ${r.symbol}`} color="#FF9500" />
+              <ResultRow label="Net" value={`${fmt(r.flow.net)} ${r.symbol}`} />
+            </>
+          )}
+
+          {r.risk?.factors?.length > 0 && (
+            <>
+              <SectionHeader title="RISK FACTORS" icon={AlertTriangle} color={riskColor} />
+              {r.risk.factors.map((f: any, i: number) => (
+                <div key={i} className="py-1.5 border-b border-[var(--border-secondary)]/20 last:border-0">
+                  <div className="flex items-center gap-2">
+                    <span className="text-[9px] font-mono font-bold" style={{ color: RISK_COLOR[f.severity] || '#00E676' }}>
+                      {f.label}
+                    </span>
+                    {f.weight > 0 && (
+                      <span className="text-[8px] font-mono text-[var(--text-muted)]">+{f.weight}</span>
+                    )}
+                  </div>
+                  <div className="text-[9px] font-mono text-[var(--text-secondary)] leading-snug mt-0.5">{f.detail}</div>
+                </div>
+              ))}
+            </>
+          )}
+
+          {r.counterparties?.length > 0 && (
+            <>
+              <SectionHeader title={`TOP COUNTERPARTIES (${r.counterparties.length})`} icon={Network} color={ACCENT} />
+              {r.counterparties.slice(0, 10).map((c: any, i: number) => (
+                <div key={i} className="flex items-center gap-2 py-1 text-[9px] font-mono">
+                  <span className={`w-[34px] flex-shrink-0 font-bold ${c.direction === 'out' ? 'text-[#FF9500]' : c.direction === 'in' ? 'text-[#00E676]' : 'text-[var(--text-muted)]'}`}>
+                    {c.direction === 'out' ? 'OUT' : c.direction === 'in' ? 'IN' : 'BOTH'}
+                  </span>
+                  <span className="flex-1 break-all text-[var(--text-primary)]">{short(c.address)}</span>
+                  <span className="text-[var(--text-muted)]">{c.txs}×</span>
+                  <span className="text-[var(--text-secondary)] w-[70px] text-right">{fmt(c.value)}</span>
+                </div>
+              ))}
+            </>
+          )}
+
+          {r.tokens?.length > 0 && (
+            <>
+              <SectionHeader title={`TOKENS (${r.tokens.length})`} icon={Layers} color={ACCENT} />
+              <div className="flex flex-wrap gap-1">
+                {r.tokens.slice(0, 20).map((t: any, i: number) => (
+                  <span key={i} className="px-1.5 py-0.5 rounded text-[9px] font-mono border border-white/15 bg-white/5 text-[var(--text-secondary)]">
+                    {t.symbol}{t.amount != null ? ` ${fmt(t.amount, 2)}` : ''}
+                  </span>
+                ))}
+              </div>
+            </>
+          )}
+
+          {r.transactions?.length > 0 && (
+            <>
+              <SectionHeader title={`RECENT TRANSACTIONS (${r.transactions.length})`} icon={Clock} color={ACCENT} />
+              {r.transactions.slice(0, 12).map((t: any, i: number) => (
+                <div key={i} className="flex items-center gap-2 py-1 text-[9px] font-mono">
+                  <span className={`w-[34px] flex-shrink-0 font-bold ${t.direction === 'out' ? 'text-[#FF9500]' : t.direction === 'in' ? 'text-[#00E676]' : 'text-[var(--text-muted)]'}`}>
+                    {t.direction === 'unknown' ? '—' : t.direction.toUpperCase()}
+                  </span>
+                  <span className="text-[var(--text-muted)] w-[64px] flex-shrink-0">{t.time ? String(t.time).slice(0, 10) : 'pending'}</span>
+                  <span className="flex-1 break-all text-[var(--text-primary)]">{short(t.hash)}</span>
+                  {t.failed && <span className="text-[#FF3D3D]">FAIL</span>}
+                  {t.value > 0 && <span className="text-[var(--text-secondary)]">{fmt(t.value)}</span>}
+                </div>
+              ))}
+            </>
+          )}
+
+          {/* Honesty rail: what this report could not establish. */}
+          {r.partial?.length > 0 && (
+            <div className="mt-3 px-2 py-1.5 rounded border border-white/10 bg-white/[0.03]">
+              <span className="text-[9px] font-mono text-[var(--text-muted)] block mb-0.5">COVERAGE LIMITS</span>
+              {r.partial.map((p: string, i: number) => (
+                <div key={i} className="text-[9px] font-mono text-[var(--text-secondary)] leading-snug">↳ {p}</div>
+              ))}
+            </div>
+          )}
+
+          <div className="mt-2 text-[8px] font-mono text-[var(--text-muted)]">
+            Sources: {(r.sources || []).join(' · ')}
+            {r.activity && ` · sampled ${r.activity.sample_size} tx${r.activity.history_complete ? ' (full history)' : ' of a longer history'}`}
+          </div>
+        </div>
+      );
+    }
+
     // ── LEAKS ──
     if (activeTab === 'leaks') {
       return (
@@ -740,6 +923,27 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
 
       {/* Input Area */}
       <div className="flex flex-col gap-1.5">
+        {/* CHAIN INTEL view switch — the brief takes no target. */}
+        {activeTab === 'crypto' && (
+          <div className="flex gap-1">
+            {([['brief', 'DAILY BRIEF'], ['wallet', 'WALLET FORENSICS']] as const).map(([id, label]) => (
+              <button
+                key={id}
+                onClick={() => setChainView(id)}
+                className="px-2 py-1 rounded text-[9px] font-mono font-bold tracking-wider transition-colors"
+                style={{
+                  color: chainView === id ? currentTab?.color : 'var(--text-muted)',
+                  background: chainView === id ? `${currentTab?.color}1a` : 'transparent',
+                  border: `1px solid ${chainView === id ? `${currentTab?.color}55` : 'rgba(255,255,255,0.1)'}`,
+                }}
+              >
+                {label}
+              </button>
+            ))}
+          </div>
+        )}
+
+        {!(activeTab === 'crypto' && chainView === 'brief') && (
         <div className="flex gap-1.5">
           <div className="flex-1 relative">
             <Search className="absolute left-2.5 top-1/2 -translate-y-1/2 w-3.5 h-3.5 text-[var(--text-muted)]" />
@@ -754,7 +958,8 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
             {loading ? <Loader2 className="w-3.5 h-3.5 animate-spin" /> : 'SCAN'}
           </button>
         </div>
-        
+        )}
+
         {/* Secondary Controls */}
         {activeTab === 'scanner' && (
           <select value={scanType} onChange={e => setScanType(e.target.value)}
@@ -1006,7 +1211,7 @@ function OsintPanelInner({ isMobile, onSweepVisualize, onScanGeolocate }: OsintP
         </div>
       )}
 
-      {results && !(sweepResult && !loading) && (
+      {(results || (activeTab === 'crypto' && chainView === 'brief')) && !(sweepResult && !loading) && (
         <div className="bg-[var(--bg-primary)]/40 border border-[var(--border-primary)] rounded-lg p-3 max-h-[50vh] overflow-y-auto styled-scrollbar">
           <div className="flex items-center justify-between mb-2">
             <span className="text-[9px] font-mono tracking-widest" style={{ color: currentTab?.color }}>{currentTab?.label} RESULTS</span>
